requirements.txt
attrs==25.1.0
boto3==1.37.12
botocore==1.37.12
cffi==1.17.1
coverage==7.6.10
cryptography==44.0.0
iniconfig==2.0.0
jmespath==1.0.1
jsonmerge==1.9.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyMySQL==1.1.1
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
s3transfer==0.11.4
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
